Unit of Work Pattern, Query enhancements and other UI touches

+++ client/src/app/components/messages/messages.component.ts
@@ -1,7 +1,8 @@
-import { Component, OnInit } from '@angular/core';
+import {Component, OnInit} from '@angular/core';
 import {Message} from "../../_models/message";
 import {Pagination} from "../../_models/pagination";
 import {MessageService} from "../../_services/message.service";
+import {ConfirmService} from "../../_services/confirm.service";
 
 @Component({
   selector: 'app-messages',
@@ -17,13 +18,15 @@ export class MessagesComponent implements OnInit {
   pageSize = 5;
   loading = false;
 
-  constructor(private messageService: MessageService) { }
+  constructor(private messageService: MessageService,
+              private ConfirmSv: ConfirmService) {
+  }
 
   ngOnInit(): void {
     this.loadMessages();
   }
 
-  loadMessages(){
+  loadMessages() {
     this.loading = true;
     this.messageService.getMessages(this.pageNumber, this.pageSize, this.container)
       .subscribe(res => {
@@ -33,16 +36,21 @@ export class MessagesComponent implements OnInit {
       });
   }
 
-  pageChanged(event: any){
-    if (this.pageNumber !== event.page){
+  pageChanged(event: any) {
+    if (this.pageNumber !== event.page) {
       this.pageNumber = event.page;
       this.loadMessages();
     }
   }
 
-  deleteMessage(id: number){
-    this.messageService.deleteMessage(id).subscribe(() => {
-      this.messages.splice(this.messages.findIndex(a => a.id === id), 1);
-    })
+  deleteMessage(id: number) {
+    this.ConfirmSv.confirm('Delete Message!', "Are you sure deleting this message?").subscribe(res => {
+      if (res) {
+        this.messageService.deleteMessage(id).subscribe(() => {
+          this.messages.splice(this.messages.findIndex(a => a.id === id), 1);
+        });
+      }
+    });
+
   }
 }
